+++ R/identify_rows.R
@@ -29,30 +29,32 @@
 #'
 #' @examples
 #' # applying to data.frame; both sepal length and width in range 3-5
-#' identify_row(iris, starts_with("Sepal"), "between", c(3, 5), if_all = TRUE)
+#' identify_rows(iris, starts_with("Sepal"), "between", c(3, 5), if_all = TRUE)
 #'
 #' # applying to remote table; species starts with se or ends with ca
 #' iris_db <- dbplyr::memdb_frame(iris)
-#' identify_row(iris_db, Species, "like", c("se%", "%ca"))
+#' identify_rows(iris_db, Species, "like", c("se%", "%ca"))
 #'
 #' # using glue_sql to write the WHERE clause
 #' # use {`vars`} to refer to the variables selected by vars
 #' # supply additional values required in the query through '...'
 #' # note that if you use LIKE here, you cannot supply multiple patterns in what
-#' identify_row(iris_db, Species, "glue_sql",
+#' identify_rows(iris_db, Species, "glue_sql",
 #'   "{`vars`} LIKE {what}",
 #'   what = "se%"
 #' )
 #'
 #' # add * after a vector
-#' identify_row(iris_db, Species, "glue_sql",
+#' identify_rows(iris_db, Species, "glue_sql",
 #'   "{`vars`} IN ({what*})",
 #'   what = c("setosa", "virginica")
 #' )
-identify_row <- function(data, vars, match = c("in", "start", "regex", "like", "between", "glue_sql"), vals, if_all = FALSE, verbose = getOption("healthdb.verbose"), query_only = TRUE, ...) {
+identify_rows <- function(data, vars, match = c("in", "start", "regex", "like", "between", "glue_sql"), vals, if_all = FALSE, verbose = getOption("healthdb.verbose"), query_only = TRUE, ...) {
   rlang::check_required(vars)
   rlang::check_required(vals)
   UseMethod("identify_rows")
 }
 
-identify_rows <- identify_row
+#' @rdname identify_rows
+#' @export
+identify_row <- identify_rows